+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # changelog
-1. Fix json: cannot unmarshal object into Go struct field Message.message.parts of type types.Part #4
+
+## unreleased
+1. Fix json: cannot unmarshal object into Go struct field Message.message.parts of type types.Part (#5)